refactor: extract duplicate match flag actions to shared module
Move identical flagMatch, updateFlag, and deleteFlag form actions into
a shared createMatchFlagActions helper. Also removes redundant
.toString() calls on reason (FormData.get already returns string).
