+++ frontend/src/routes/(authed)/player/[id]/+page.server.ts
@@ -1,4 +1,5 @@
 import type { MatchTeamV2 } from '../../../../api';
+import { createMatchFlagActions } from '$lib/server/actions/matchFlagActions';
 import type { Actions, PageServerLoad } from './$types';
 import type { MemberLeaderboardEntry } from './types';
 import { movePlayerToMember1 } from './utils';
@@ -168,85 +169,15 @@ const millisecondsToDays = (ms: number) => Math.round(ms / 1000 / 60 / 60 / 24);
 
 export const actions: Actions = {
   flagMatch: async ({ request, locals: { apiClient } }) => {
-    const data = await request.formData();
-    const matchId = data.get('matchId');
-    const reason = data.get('reason');
-
-    if (!matchId || !reason) {
-      return fail(400, { success: false, message: 'Match ID and reason are required' });
-    }
-
-    const parsedMatchId = Number(matchId);
-    if (!Number.isInteger(parsedMatchId) || parsedMatchId <= 0) {
-      return fail(400, { success: false, message: 'Valid match ID is required' });
-    }
-
-    try {
-      await apiClient.matchFlagsApi.matchFlagsCreateFlag({
-        createMatchFlagRequest: {
-          matchId: parsedMatchId,
-          reason: reason.toString(),
-        },
-      });
-
-      return { success: true, message: 'Match flagged successfully' };
-    } catch (error) {
-      console.error('Error flagging match:', error);
-      return fail(500, { success: false, message: 'Failed to flag match' });
-    }
+    const actions = createMatchFlagActions(apiClient);
+    return actions.flagMatch(await request.formData());
   },
-
   updateFlag: async ({ request, locals: { apiClient } }) => {
-    const data = await request.formData();
-    const flagId = data.get('flagId');
-    const reason = data.get('reason');
-
-    if (!flagId || !reason) {
-      return fail(400, { success: false, message: 'Flag ID and reason are required' });
-    }
-
-    const parsedFlagId = Number(flagId);
-    if (!Number.isInteger(parsedFlagId) || parsedFlagId <= 0) {
-      return fail(400, { success: false, message: 'Valid flag ID is required' });
-    }
-
-    try {
-      await apiClient.matchFlagsApi.matchFlagsUpdateFlag({
-        id: parsedFlagId,
-        updateMatchFlagReasonRequest: {
-          reason: reason.toString(),
-        },
-      });
-
-      return { success: true, message: 'Flag updated successfully' };
-    } catch (error) {
-      console.error('Error updating flag:', error);
-      return fail(500, { success: false, message: 'Failed to update flag' });
-    }
+    const actions = createMatchFlagActions(apiClient);
+    return actions.updateFlag(await request.formData());
   },
-
   deleteFlag: async ({ request, locals: { apiClient } }) => {
-    const data = await request.formData();
-    const flagId = data.get('flagId');
-
-    if (!flagId) {
-      return fail(400, { success: false, message: 'Flag ID is required' });
-    }
-
-    const parsedFlagId = Number(flagId);
-    if (!Number.isInteger(parsedFlagId) || parsedFlagId <= 0) {
-      return fail(400, { success: false, message: 'Valid flag ID is required' });
-    }
-
-    try {
-      await apiClient.matchFlagsApi.matchFlagsDeleteFlag({
-        id: parsedFlagId,
-      });
-
-      return { success: true, message: 'Flag deleted successfully' };
-    } catch (error) {
-      console.error('Error deleting flag:', error);
-      return fail(500, { success: false, message: 'Failed to delete flag' });
-    }
+    const actions = createMatchFlagActions(apiClient);
+    return actions.deleteFlag(await request.formData());
   },
 };
